separate messaging from command
